+++ libs/constants.py
@@ -1,3 +1,4 @@
+APP_NAME = 'BEAGLES'
 SETTING_FILENAME = 'filename'
 SETTING_RECENT_FILES = 'recentFiles'
 SETTING_WIN_SIZE = 'window/size'
@@ -16,3 +17,4 @@ FORMAT_PASCALVOC='PascalVOC'
 FORMAT_YOLO='YOLO'
 SETTING_DRAW_SQUARE = 'draw/square'
 DEFAULT_ENCODING = 'utf-8'
+
